+++ HISTORY.rst
@@ -8,7 +8,9 @@ History
 * Add "naive" and "top-k" methods in MapieClassifier
 * Include J+aB method in regression tutorial
 * Add MNIST example for classification
-* Add replication of the Chen Xu's tutorial testing Jackknife+aB vs Jackknife+.
+* Uniformize the use of matrix k_ and add an argument "ensemble" to method 
+"predict" in regression.py
+* Add replication of the Chen Xu's tutorial testing Jackknife+aB vs Jackknife+
 * Add Jackknife+-after-Bootstrap documentation
 
 0.3.1 (2021-11-19)